Move `rootContext` method to Run from Compiler
